Added ExpressionSolver class for solving math expressions.
_find_highest_precedence finds the index of the operator with the highest precedence.
_execute_operation calculates the result of a single operator based on its type
(INFIX, PREFIX, or POSTFIX).
_replace_with_result replaces the evaluated sub-expression with its result in the tokens list.
solve_parentheses resolves both regular and nested parentheses recursively.
solve runs the main evaluation loop until a single result remains.

Also changed precedence attribute to float and set unary minus precedence to 2.5.

+++ operator.py
@@ -3,7 +3,7 @@ from enums import OperatorType
 
 
 class Operator(ABC):
-    def __init__(self, symbol: str, precedence: int, operator_type: OperatorType):
+    def __init__(self, symbol: str, precedence: float, operator_type: OperatorType):
         self.symbol = symbol
         self.precedence = precedence
         self.operator_type = operator_type
